adds support for StravaResultKit
