+++ Sources/StravaUpload.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-internal enum UploadResourcePath: String {
+public enum UploadResourcePath: String {
     case Upload = "/v3/uploads"
     case CheckUpload = "/api/v3/uploads/:id"
 }
